+++ tests/message_properties/definitions/test_labware.py
@@ -1,4 +1,4 @@
-from tol_lab_share.messages.properties import Value
+from tol_lab_share.messages.properties.value_wrappers import Value
 from tol_lab_share.message_properties.definitions.labware import Labware
 from tol_lab_share.message_properties.definitions.location import Location
 from tol_lab_share.messages.traction import TractionReceptionMessage, TractionQcMessage
