+++ tests/message_properties/definitions/test_labware_type.py
@@ -1,5 +1,5 @@
 from tol_lab_share.message_properties.definitions.labware_type import LabwareType
-from tol_lab_share.messages.properties import Value
+from tol_lab_share.messages.properties.value_wrappers import Value
 
 
 def test_LabwareType_check_LabwareType_is_string():
