+++ tests/message_properties/definitions/test_scientific_name_from_taxon_id.py
@@ -1,5 +1,5 @@
 from tol_lab_share.message_properties.definitions.scientific_name_from_taxon_id import ScientificNameFromTaxonId
-from tol_lab_share.messages.properties import Value
+from tol_lab_share.messages.properties.value_wrappers import Value
 from tol_lab_share.error_codes import ExceptionErrorCode
 from helpers import check_validates_integer_string
 import requests_mock
