+++ tests/message_properties/definitions/test_uuid.py
@@ -1,5 +1,5 @@
 from tol_lab_share.message_properties.definitions.uuid import Uuid
-from tol_lab_share.messages.properties import Value
+from tol_lab_share.messages.properties.value_wrappers import Value
 
 
 def test_uuid_check_is_uuid_detects_invalid_uuid():
