+++ tests/messages/properties/value_wrappers/test_dict_value.py
@@ -1,4 +1,4 @@
-from tol_lab_share.messages.properties import DictValue
+from tol_lab_share.messages.properties.value_wrappers import DictValue
 
 
 def test_dict_input_can_validate_when_valid():
