+++ tol_lab_share/messages/properties/__init__.py
@@ -1,3 +1,1 @@
-from .dict_value import DictValue
 from .message_property import MessageProperty
-from .value import Value
