+++ tol_lab_share/messages/properties/value_wrappers/__init__.py
@@ -1,0 +1,2 @@
+from .dict_value import DictValue
+from .value import Value
