+++ tol_lab_share/messages/traction/qc_message.py
@@ -9,7 +9,7 @@ from tol_lab_share import error_codes
 from tol_lab_share.constants import OUTPUT_TRACTION_MESSAGE_SOURCE
 from tol_lab_share.error_codes import ErrorCode
 from tol_lab_share.helpers import get_config
-from tol_lab_share.messages.properties import Value
+from tol_lab_share.messages.properties.value_wrappers import Value
 from tol_lab_share.messages.properties.message_property import MessageProperty
 from tol_lab_share.messages.output_feedback_message import OutputFeedbackMessage
